clean up
code cleaning
