+++ src/char_buffer.c
@@ -12,11 +12,9 @@
 
 #include "char_buffer.h"
 
-#define MIN_LENGTH 1024
-
-// place holder ....
 void* connectionSSL = NULL;
 
+#define MIN_LENGTH 1024
 
 cb_t cb_new() {
   cb_t self = (cb_t)malloc(sizeof(struct cb_struct));
