+++ src/webinspector.c
@@ -40,6 +40,8 @@
 // some arbitrarly limit, to catch bad packets
 #define MAX_BODY_LENGTH 1<<26
 
+extern idevice_connection_t connectionSSL;
+
 struct wi_private {
   bool partials_supported;
   cb_t in;
@@ -48,33 +50,11 @@ struct wi_private {
   size_t body_length;
 };
 
-// iOS 13 --------------------------------------
-int g_vers[3] = {0, 0, 0};
-
- struct service_client_private 
- {
- 	idevice_connection_t connection;
- };
-	
-extern idevice_connection_t connectionSSL;
-// iOS 13 --------------------------------------
- 
-
 //
 // CONNECT
 //
 
 #ifndef HAVE_IDEVICE_CONNECTION_GET_FD
-// based on libimobiledevice/src/idevice.h
-enum connection_type {
-  CONNECTION_USBMUXD = 1
-};
-struct idevice_connection_private {
-  char *udid;  // added in v1.1.6
-  enum connection_type type;
-  void *data;
-  void *ssl_data;
-};
 
 wi_status idevice_connection_get_fd(idevice_connection_t connection, int *to_fd) 
 {
@@ -83,26 +63,13 @@ wi_status idevice_connection_get_fd(idevice_connection_t connection, int *to_fd)
   
   idevice_connection_private *c = ((sizeof(*connection) == sizeof(idevice_connection_private)) ? (idevice_connection_private *) connection : NULL);
   
-  bool bInvalidStructure = false;
-  
-  if(g_vers[0] >= 13)
-  {
-  	 if (!c || c->type != CONNECTION_USBMUXD || c->data <=0) 
-		 bInvalidStructure = true;
-  }
-  else if (!c || c->type != CONNECTION_USBMUXD || c->data <= 0 || c->ssl_data) 
-  {
-	  bInvalidStructure = true;
-  }
-  
-  if(bInvalidStructure)
+  if (!c || c->type != CONNECTION_USBMUXD || c->data <= 0 || c->ssl_data) 
   {
       perror("Invalid idevice_connection struct.  Please verify that "
           __FILE__ "'s idevice_connection_private matches your version of"
           " libimbiledevice/src/idevice.h");
       return WI_ERROR;
   }
-  
   
   int fd = (int)(long)c->data;
   struct stat fd_stat;
@@ -115,20 +82,6 @@ wi_status idevice_connection_get_fd(idevice_connection_t connection, int *to_fd)
 }
 #endif
 
-// ------- MICKEL
-enum connection_type {
-	CONNECTION_USBMUXD = 1
-};
-struct idevice_connection_private {
-	char *udid;  // added in v1.1.6
-	enum connection_type type;
-	void *data;
-	void *ssl_data;
-};
-// ------- MICKEL
-
-
-
 int wi_connect(const char *device_id, char **to_device_id,
     char **to_device_name, int *to_device_os_version, int recv_timeout) {
   int ret = -1;
@@ -139,6 +92,7 @@ int wi_connect(const char *device_id, char **to_device_id,
   lockdownd_client_t client = NULL;
   idevice_connection_t connection = NULL;
   int fd = -1;
+  int vers[3] = {0, 0, 0};
 
   // get phone
   if (idevice_new(&phone, device_id)) {
@@ -173,10 +127,10 @@ int wi_connect(const char *device_id, char **to_device_id,
     char *s_version = NULL;
     plist_get_string_val(node, &s_version);
     if (s_version && sscanf(s_version, "%d.%d.%d",
-          &g_vers[0], &g_vers[1], &g_vers[2]) >= 2) {
-      *to_device_os_version = ((g_vers[0] & 0xFF) << 16) |
-                              ((g_vers[1] & 0xFF) << 8)  |
-                               (g_vers[2] & 0xFF);
+          &vers[0], &vers[1], &vers[2]) >= 2) {
+      *to_device_os_version = ((vers[0] & 0xFF) << 16) |
+                              ((vers[1] & 0xFF) << 8)  |
+                               (vers[2] & 0xFF);
     } else {
       *to_device_os_version = 0;
     }
@@ -197,23 +151,16 @@ int wi_connect(const char *device_id, char **to_device_id,
     goto leave_cleanup;
   }
 
+  if(vers[0]>=13) {
+	  service_client_t client_srv = (service_client_t)malloc(sizeof(struct service_client_private));
+	  client_srv->connection = connection;
 
-  // iOS 13.x ---------------------------------------------------------------------
-  if(g_vers[0]>=13) // the wi connection is ssl based started iOS 13.0 and higher ...
-  {
-	  	service_client_t client_srv = (service_client_t)malloc(sizeof(struct service_client_private));
-
-	  	  client_srv->connection = connection;
-
-	  	  /* enable SSL if requested */
-	  	  if (service->ssl_enabled == 1)
-		  {
+	  /* enable SSL if requested */
+	  if (service->ssl_enabled == 1){
 	  		    service_enable_ssl(client_srv);
 		  		connectionSSL = client_srv->connection;
 	  	  }
   }
-
- // iOS 13.x ---------------------------------------------------------------------
 
   if (client) {
     // not needed anymore
